+++ src/enforce-limit.js
@@ -1,88 +1,5 @@
-const tokenize = require("./tokenize");
 const createToken = require("./create-token");
-
-/**
- * Splits sql text into an array of arrays of tokens
- * @param {string} sqlText
- */
-function getQueriesTokens(sqlText) {
-  const tokens = tokenize(sqlText);
-  const queriesTokens = [];
-  let queryTokens = [];
-  tokens.forEach((token) => {
-    queryTokens.push(token);
-    if (token.type === "terminator") {
-      queriesTokens.push(queryTokens);
-      queryTokens = [];
-    }
-  });
-  // push last set
-  if (queryTokens.length) {
-    queriesTokens.push(queryTokens);
-  }
-
-  return queriesTokens;
-}
-
-/**
- * Find if a query is a select query or not.
- * CTEs *may* be a select. They could also be inserts/updates
- * @param {array<object>} queryTokens
- */
-function getStatementType(queryTokens = []) {
-  // Queries could be wrapped in parens (WITH something AS (some query) SELECT 'hello' AS world;)
-  // In event that this query wrapped in parents, this needs to track what level of parens need the limit injected into
-  let parenLevel = 0;
-  let targetParenLevel;
-  // statementKeyword will be `select`, `insert`, `alter`, etc.
-  // keywords `with` and `as` not included to filter out cte
-  let statementKeyword = "";
-  let statementkeywordIndex;
-
-  for (let index = 0; index < queryTokens.length; index++) {
-    const token = queryTokens[index];
-    if (token.type === "lparen") {
-      parenLevel++;
-    } else if (token.type === "rparen") {
-      parenLevel--;
-    } else if (token.type === "keyword") {
-      // If targetParenLevel has not yet been set,
-      // we are dealing with the first keyword, which informs us of the "level"
-      // we want to consider for finding SELECT and TOP/LIMIT
-      if (targetParenLevel === undefined) {
-        targetParenLevel = parenLevel;
-      }
-      // Statement keyword we are considering not something found in prep of CTE
-      // If the current keyword isn't a `with` and not `as`, and at the same level as our targetParenLevel,
-      // We can assume it tells us what kind of query we are dealing with.
-      // Consider queries like the following queries
-      //
-      // WITH cte AS (...) SELECT ...
-      // WITH cte AS (...) INSERT INTO ... SELECT
-      // WITH cte AS (...) UPDATE ... FROM ...
-      // (WITH cte AS (...) SELECT ...)
-      if (
-        !statementKeyword &&
-        targetParenLevel === parenLevel &&
-        token.value !== "with" &&
-        token.value !== "as"
-      ) {
-        statementKeyword = token.value;
-        statementkeywordIndex = index;
-
-        // We've identified the statement keyword
-        // We can exit the loop
-        index = queryTokens.length;
-      }
-    }
-  }
-
-  return {
-    statementkeywordIndex,
-    statementKeyword,
-    targetParenLevel,
-  };
-}
+const getStatementType = require("./get-statement-type");
 
 function findParenLevelToken(tokens, startingIndex, predicate) {
   let level = 0;
@@ -118,7 +35,7 @@ function nextNonCommentNonWhitespace(tokens, startingIndex) {
 }
 
 function hasLimit(tokens, startingIndex) {
-  const limitKeywordToken = findToken(
+  const limitKeywordToken = findParenLevelToken(
     tokens,
     startingIndex,
     (token) => token.type === "keyword" && token.value === "limit"
@@ -261,42 +178,4 @@ function enforceLimit(queryTokens, limit) {
   ];
 }
 
-/**
- * Takes SQL text and splits it by terminator
- * Returns array of single SQL statements.
- * Extra spaces trailing terminator are not returned for a query
- * @param {string} sqlText
- * @param {boolean} includeTerminators = include terminators in individual queries
- */
-function getQueries(sqlText, includeTerminators = false) {
-  const queries = [];
-  const queriesTokens = getQueriesTokens(sqlText);
-  queriesTokens.forEach((queryTokens) => {
-    // if set of tokens has something other than whitespace/terminators, consider it
-    if (
-      queryTokens.filter(
-        (t) => t.type !== "whitespace" && t.type !== "terminator"
-      ).length > 0
-    ) {
-      if (includeTerminators) {
-        const query = queryTokens.map((token) => token.text).join("");
-        queries.push(query);
-      } else {
-        const query = queryTokens
-          .filter((token) => token.type !== "terminator")
-          .map((token) => token.text)
-          .join("");
-        queries.push(query);
-      }
-    }
-  });
-  return queries;
-}
-
-module.exports = {
-  enforceLimit,
-  getQueries,
-  getQueriesTokens,
-  getStatementType,
-  tokenize,
-};
+module.exports = enforceLimit;
