+++ src/get-queries-tokens.js
@@ -1,0 +1,26 @@
+const tokenize = require("./tokenize");
+
+/**
+ * Splits sql text into an array of arrays of tokens
+ * @param {string} sqlText
+ */
+function getQueriesTokens(sqlText) {
+  const tokens = tokenize(sqlText);
+  const queriesTokens = [];
+  let queryTokens = [];
+  tokens.forEach((token) => {
+    queryTokens.push(token);
+    if (token.type === "terminator") {
+      queriesTokens.push(queryTokens);
+      queryTokens = [];
+    }
+  });
+  // push last set
+  if (queryTokens.length) {
+    queriesTokens.push(queryTokens);
+  }
+
+  return queriesTokens;
+}
+
+module.exports = getQueriesTokens;
