+++ src/get-statement-type.js
@@ -1,0 +1,61 @@
+/**
+ * Find if a query is a select query or not.
+ * CTEs *may* be a select. They could also be inserts/updates
+ * @param {array<object>} queryTokens
+ */
+function getStatementType(queryTokens = []) {
+  // Queries could be wrapped in parens (WITH something AS (some query) SELECT 'hello' AS world;)
+  // In event that this query wrapped in parents, this needs to track what level of parens need the limit injected into
+  let parenLevel = 0;
+  let targetParenLevel;
+  // statementKeyword will be `select`, `insert`, `alter`, etc.
+  // keywords `with` and `as` not included to filter out cte
+  let statementKeyword = "";
+  let statementkeywordIndex;
+
+  for (let index = 0; index < queryTokens.length; index++) {
+    const token = queryTokens[index];
+    if (token.type === "lparen") {
+      parenLevel++;
+    } else if (token.type === "rparen") {
+      parenLevel--;
+    } else if (token.type === "keyword") {
+      // If targetParenLevel has not yet been set,
+      // we are dealing with the first keyword, which informs us of the "level"
+      // we want to consider for finding SELECT and TOP/LIMIT
+      if (targetParenLevel === undefined) {
+        targetParenLevel = parenLevel;
+      }
+      // Statement keyword we are considering not something found in prep of CTE
+      // If the current keyword isn't a `with` and not `as`, and at the same level as our targetParenLevel,
+      // We can assume it tells us what kind of query we are dealing with.
+      // Consider queries like the following queries
+      //
+      // WITH cte AS (...) SELECT ...
+      // WITH cte AS (...) INSERT INTO ... SELECT
+      // WITH cte AS (...) UPDATE ... FROM ...
+      // (WITH cte AS (...) SELECT ...)
+      if (
+        !statementKeyword &&
+        targetParenLevel === parenLevel &&
+        token.value !== "with" &&
+        token.value !== "as"
+      ) {
+        statementKeyword = token.value;
+        statementkeywordIndex = index;
+
+        // We've identified the statement keyword
+        // We can exit the loop
+        index = queryTokens.length;
+      }
+    }
+  }
+
+  return {
+    statementkeywordIndex,
+    statementKeyword,
+    targetParenLevel,
+  };
+}
+
+module.exports = getStatementType;
